# backend-v2/src/routes/puzzles.ts
/*
 * ⚠️  WARNING: GENERATED CODE - DO NOT MODIFY ⚠️
 * 
 * This file is automatically generated by the backend code generator.
 * Any manual changes will be overwritten when the generator runs.
 * 
 * To make changes, modify the generator configuration files in:
 * /tools/backend-tools/backend-generator-v2/
 * 
 * Then re-run the generator to update this file.
 */

import { Router } from 'express';
import { authenticate } from '../middleware/auth';
import { validate } from '../middleware/validation';
import { PuzzleService } from '../services/puzzlesService';
import { PuzzleResponse, CreatePuzzleRequest, UpdatePuzzleRequest } from '../models/Puzzle';

const router = Router();
const puzzlesService = new PuzzleService();

// POST /
router.post('/', authenticate, async (req: any, res) => {
  try {
    const result = await puzzlesService.createPuzzle(req.body);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// GET /:id
router.get('/:id', authenticate, async (req: any, res) => {
  try {
    const result = await puzzlesService.getPuzzleById(req.params.id);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// PUT /:id
router.put('/:id', authenticate, async (req: any, res) => {
  try {
    const result = await puzzlesService.updatePuzzle(req.params.id, req.body);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// DELETE /:id
router.delete('/:id', authenticate, async (req: any, res) => {
  try {
    const result = await puzzlesService.deletePuzzle(req.params.id);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// GET /
router.get('/', authenticate, async (req: any, res) => {
  try {
    const result = await puzzlesService.listPuzzles();
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

export default router;